Fixed confusion between UsersSeeder and UserRecordIssuersSeeder
